Rename _auth.js to auth.js
